requirements.txt
flask
pytest
python-dotenv
fasteners